requirements.txt.txt
requests==2.32.3
beautifulsoup4==4.12.3
python-telegram-bot==20.9
pytz==2024.1